requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
onnx==1.19.1
onnxslim
onnxruntime-gpu>=1.15.0
tensorrt>=8.6.0
pycuda>=2022.1
albumentations>=1.3.0
opencv-python>=4.8.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart
scikit-learn
matplotlib
seaborn
nvidia-ml-py
lapx
scipy
pytest
requests
